+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentrySubsystem.cpp
@@ -672,6 +672,7 @@ void FGenericPlatformSentrySubsystem::InitWithSettings(const USentrySettings* se
 	{
 		sentry_options_set_minidump_mode(options, FGenericPlatformSentryConverters::MinidumpModeToNative(settings->MinidumpMode));
 		sentry_options_set_crash_reporting_mode(options, FGenericPlatformSentryConverters::CrashReportingModeToNative(settings->CrashReportingMode));
+		sentry_options_set_thread_stackwalk_mode(options, FGenericPlatformSentryConverters::ThreadStackwalkModeToNative(settings->ThreadStackwalkMode));
 		sentry_options_set_crash_upload_mode(options, FGenericPlatformSentryConverters::CrashUploadModeToNative(settings->CrashUploadMode));
 	}